Move Calico's mechanism driver to networking-calico
The Calico mechanism driver was previously part of
https://github.com/projectcalico/calico, but logically belongs in
networking-calico - the OpenStack/Neutron integration code for Calico
- and so is now moving to this repository.  As well as the mechanism
driver code, this move includes:

- unit testing for the mechanism driver

- creation of a neutron.ml2.mechanism_drivers entry point for the name
  'calico'

- requirements updates as needed for the new code.

Also, so that Calico's Debian and RPM package-based installations
continue to work, and so that existing installations can be smoothly
upgraded, this change adds Debian and RPM packaging to
networking-calico.  Established Calico installation instructions [1,2]
use the umbrella package names 'calico-control' and 'calico-compute'
to indicate whatever is needed to add Calico function to an OpenStack
controller and compute node respectively, and (as these are
OpenStack-specific concepts) responsibility for generating these
umbrella packages also now moves from
https://github.com/projectcalico/calico to networking-calico.

calico-control and calico-compute both depend on a third package,
'networking-calico', which is the package that contains
networking-calico's Python code.  Thus an install of the
'networking-calico' Debian or RPM package should be very similar to a
'pip install networking-calico' from PyPI.

[1] http://docs.projectcalico.org/en/latest/ubuntu-opens-install.html
[2] http://docs.projectcalico.org/en/latest/redhat-opens-install.html

Change-Id: I54724a4e4fbb465daddcff07e6da19cb91b2dc4a