Buffer test logging instead of printing it, to save performance
For some reason, test code 'print'ing adds 10 or 20 seconds to each
test suite run, even though we never actually see the printed
information (because tox or one of the other test runner layers
captures it).

So, let's not do that.  But, the existing test code has a lot of
logging-like statements that could well be useful in understanding and
debugging some future test situation.  Therefore we convert each
'print' to a 'testlog' call, which just adds to an internal array.
Then in future, if a developer wants to see that buffer at a
particular point, they can insert a call to lib.printtestlogs().

Incidentally this new arrangement also allows us to verify that the
performance cost is associated specifically with printing.  If I add
lib.printtestlogs() to test_noop_entry_points - a test that normally
takes no time - that test takes 10 seconds!

(At a guess, eventlet or testr is monkey-patching sys.stdout and
writes are blocking for ~300ms on some print calls.  Possibly some
interaction with our mocking out of time and eventlets monkey
patching.

Found by logging timstamps and doing a binary chop!)

Change-Id: I651588bdcf4a35980a7b82933cf9de6a1abd10b4