Comment-out print statements.
Not a "real" solution but thought I'd put this up as a review job
so it doesn't get lost.

I'm not sure why but the test that was taking 30s runs in 8s if
I remove the print statements.

At a guess, eventlet or testr is monkey-patching sys.stdout and
writes are blocking for ~300ms on some print calls.  Possibly
some interaction with our mocking out of time and eventlets monkey
patching.

(Found by logging timstamps and doing a binary chop!)

Change-Id: I651588bdcf4a35980a7b82933cf9de6a1abd10b4