New DHCP agent driven by etcd data instead of by Neutron RPC
A Calico network is not bridged across all the involved compute hosts,
and therefore requires a different DHCP provision approach from the
standard Neutron ones, which provide DHCP for all instances on a
Neutron Network from a single DHCP agent located on the 'network
node'.  (Or from a few DHCP agents, for HA.)  For Calico thus far, the
approach we've taken is to run a DHCP agent on each compute host.

It turns out, though, that (with more than about a couple of hundred
compute hosts) the agent/server RPC communications place too much load
on the Neutron server(s), such that the Neutron servers become
unresponsive to Nova when the latter is trying to bring up a new VM;
and hence new VM creations fail.  For more details, see the referenced
bug.

In Calico we already have another mechanism for communicating between
the controller and compute nodes - the distributed etcd database - and
this mechanism already includes almost all the information that is
needed for DHCP provision.  Therefore we can avoid the problem in
scaling of the Neutron RPC mechanism by creating and using an
etcd-driven DHCP agent instead.

This commit contains the following consequent changes.

- Add a Calico-specific top-level DHCP agent class and executable
  script, like the Neutron analogs but driven by etcd data instead of
  by Neutron RPC.  (For more detailed design see
  doc/source/dhcp-agent.rst.)

- Update the Calico DevStack plugin so that it runs the Calico DHCP
  agent instead of the reference Neutron DHCP agent.

- Revert a recently added optimization in
  networking_calico/agent/linux/dhcp.py that limits the config that is
  fed to Dnsmasq to be only what is needed for the local compute host
  (as opposed to for all ports on all compute hosts).  This is now not
  needed, as the etcd mechanism already limits the information that it
  processes to what is relevant on the local compute host.

- Enhance the Calico mechanism driver so that it writes per-subnet
  information, that the DHCP agent needs, into etcd.  Specifically
  this means, for each subnet:

  - the subnet CIDR

  - any DNS servers that are configured for the subnet.

  Previously there was no subnet-scope data in Calico's etcd data
  model.

- Enhance the Calico mechanism driver so that it writes additional
  information that the DHCP agent needs for each endpoint, namely:

  - for each fixed IP, which subnet it is associated with

  - the FQDN that has been configured for the endpoint, if any.

- Related new and updated testing.

Closes-Bug: #1519803
Change-Id: Ia4ea03cf0b006652ecd0a718842104a6404273f1