Prefix the profile IDs that we program into etcd
This change adds an OpenStack-specific prefix to the ID of each security
profile that this driver programs into etcd (the shared Calico
database), so that profiles programmed by this driver can happily
coexist with those that might be programmed by other Calico
orchestrators.  Specifically, it means that this driver now will not
accidentally delete profiles from other orchestrators.

This change also improves the ordering of how it writes profile and
endpoint data into etcd, so that it should never write an endpoint that
references a profile that isn't there yet, and should never delete a
profile while there are still endpoints that reference it.  This should
eliminate possible, if small, windows, in which Felix sees an endpoint
with a profile that is not defined in etcd.

Change-Id: I2c1ce521b59269bb07decb98449c68717182589d
Sem-Ver: bugfix