Change user for calico-dhcp-agent
By default the calico dhcp agent want's to run as root. A lot of trouble is awoided if the agent runs as the neutron user, like the neutron dhcp agent. These changes accomplishes that by changing the user in the systemd unit file and ensures that the logging directory is present with correct permissions.

This creates a dependency to openstack-neutron-common, since that package creates the neutron user and installs the necessary rootwrap scripts.

Change-Id: Ie7124ab11d3ad3b94b8019caf45e7e95a7517545