DHCP agent log to /var/log/neutron instead of .../calico
The Calico DHCP agent runs more like a Neutron component than a Calico
component in the sense that it uses a lot of Neutron library code, which
includes two specific things:

1. It uses Neutron's file-based locking mechanism, which requires
   accessing the lock directory with similar credentials as other
   Neutron components.

2. It uses rootwrap when it needs to do privileged operations, and so
   does not need as a whole to run as root.

Therefore it makes sense for the Calico DHCP agent to run as a
non-privileged user.  The DevStack plugin already runs it as such, and
patches are up to modify the package-based installs of networking-calico
to run the DHCP agent as 'neutron' instead of as root.

It also then follows that the DHCP agent should log to /var/log/neutron
- sharing that directory with other Neutron components that log -
instead of to /var/log/calico.  With /var/log/calico, in a DevStack
install, there is a race between calico-felix and calico-dhcp-agent: if
calico-felix tries to create the directory first, it will not be
writable by calico-dhcp-agent, and then the DHCP will die.

Change-Id: I63db16e93ce1ac02ceb955fbbd40aae10f16a690