DevStack: make /var/log/calico writable by DHCP agent
In a DevStack install, the Calico DHCP agent runs as an unprivileged
user, instead of as root.  This is appropriate because it uses a lot of
Neutron library code, which includes two specific things:

1. It uses Neutron's file-based locking mechanism, which requires
   accessing the lock directory with similar credentials as other
   Neutron components.

2. It uses rootwrap when it needs to do privileged operations, and so
   does not need as a whole to run as root.

(Correspondingly, patches are currently up to modify the package-based
installs of networking-calico to run the DHCP agent as 'neutron' instead
of as root.)

The log directory /var/log/calico is created by calico-felix, which runs
as root, and so needs to be chmod'd to allow calico-dhcp-agent to write
to it too.

Change-Id: If85e27ad12df8ec35d3723ce316889a13124e659