Fix IPv6 router advertisements with multiple networks
This involves a significant rework of the Calico DHCP agent,
specifically to run an instance of Dnsmasq for each Neutron network that
is in use on the compute host where the DHCP agent is running (as the
reference Neutron DHCP agent does); instead of running just a single
instance of Dnsmasq for all of the Neutron networks that are in use on
the compute host.

The Calico DHCP agent has previously been using (what it thought was) an
optimization / simplification compared to the reference Neutron DHCP
agent: it runs a single Dnsmasq instance provisioned with the collated
subnet information for all Neutron networks that are relevant to
instances hosted by the local compute node; whereas the reference
Neutron DHCP agent runs a Dnsmasq instance _per Neutron network_,
provisioned only with the subnet information for that network.

Calico's simplification works correctly for DHCPv4 and DHCPv6, but
unfortunately not for IPv6 Router Advertisements (RAs): the effect of
the simplification is that the same RA, including router prefixes for
all of the Neutron networks that some local instances use, is sent to
every local TAP interface.  This means that:

- a port that is attached to a network without any IPv6 subnets will
  still get IPv6 routing information

- if different local ports are attached to networks with different
  IPv6 subnets, they will receive each other's routing information as
  well as their own.

Therefore this commit removes that simplification, and now the Calico
DHCP agent will run a separate Dnsmasq instance for each relevant
Neutron network.

The main constituents of this change are:

- adding each subnet's network ID to the information about that subnet
  that the Calico ML2 driver stores in etcd

- populating and maintaining the NetworkCache with multiple NetModels,
  reflecting the true Neutron network/subnet structure, instead of with
  a single collated NetModel

- instructing each Dnsmasq to listen on precisely the relevant TAP
  interfaces, instead of on a wildcard ('tap*').

Sem-Ver: bugfix
Change-Id: I3f2e9cab2c109402426a3cc73612360463767bf7