Correctly handle etcd being stopped
Calico uses etcd as a distributed database, so the Calico mechanism
driver relies on connection to etcd (either directly to an etcd master
node, or to an etcd proxy).  In an HA setup, the Calico driver uses
etcd (as well as for other purposes) to determine which driver instance
is the current leader; and one of the responsibilities of the leader
instance is to propagate Calico agent status from etcd to the Neutron
DB.

If the driver's connection to etcd is interrupted - e.g. by stopping the
local etcd proxy - the thread monitoring leader status incorrectly
exits, and is not recreated until the driver is asked to handle a new
network/subnet/port CRUD
operation (ref. https://github.com/projectcalico/calico/issues/543),
even if etcd is started again quite quickly.  Therefore, if there is
only one Neutron server - i.e. no HA - after a short period the Calico
agent up status will time out and Calico agents will be considered to be
down, until the next CRUD operation recreates the leader status thread.

Often the next operation is a port create (for a new instance) and this
will fail because of there appearing to be no Calico agents available.

The root problem here is the leader status thread exiting, which is in
turn caused by a requests/urllib3 unvendoring issue that prevents
python-etcd from catching exceptions raised by urllib3.  This commit
fixes that problem, so the leader status thread doesn't exit, and Calico
agents are reported as up again within a few seconds after etcd has been
restarted.

Change-Id: Icf037062be401289f2c514e278e61d6714b27119
Sem-Ver: bugfix