Import needed code from core Calico repo
networking-calico currently depends on a separate Calico repo[1] that
contains (a) the Calico agent, Felix and (b) utility code that is
independent of OpenStack or of the other systems that Calico integrates
with.  However a couple of things are happening to that repo.  Firstly
it's being renamed as 'felix', and secondly much of it is being
rewritten in Go.  Therefore it makes sense to move the Python utility
code that networking-calico needs from that repo, into
networking-calico.  This also simplifies networking-calico's
dependencies.

[1] https://github.com/projectcalico/calico/, soon to be renamed as
https://github.com/projectcalico/felix/.

Change-Id: I99d1bb919ab3119826d1b410e9ccaa4ba2f35696