Capture Felix logging in OpenStack CI runs
The OpenStack CI process (devstack-gate) saves off all logs generated by
systemd services that are started with "run_process", but not _all_
services, and so not calico-felix.  But to get the logging from
calico-felix we can use "run_process" to start a secondary service that
just tails Felix's log.  That's what we do here, and the Felix log is
then available in captured logs as a file named
screen-felix-logger.txt.gz.

Change-Id: I40957a3ed4f96a68272a50bcc40303dde770553f