Fixes needed to networking-calico etcdv3 usage
This commit contains 4 fixes that are all related to
networking-calico's etcdv3 usage, and that have been revealed through CI
testing.  They are in detail unrelated, but I think it makes sense to
commit them together, as they are all needed ASAP as fixups to
https://review.openstack.org/#/c/544586/.

1. Make DHCP agent handle etcd being initially unavailable

Specifically this is to handle errors like this, when etcd is not yet
available when the DHCP agent starts:

2018-02-27 23:43:35,846 [39149232] (CRITICAL) log: Unhandled error
2018-02-27 23:43:35.846 18623 ERROR neutron Traceback (most recent call last):
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/bin/calico-dhcp-agent", line 10, in <module>
2018-02-27 23:43:35.846 18623 ERROR neutron     sys.exit(main())
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/networking_calico/agent/dhcp_agent.py", line 611, in main
2018-02-27 23:43:35.846 18623 ERROR neutron     agent.run()
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/osprofiler/profiler.py", line 153, in wrapper
2018-02-27 23:43:35.846 18623 ERROR neutron     return f(*args, **kwargs)
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/networking_calico/agent/dhcp_agent.py", line 577, in run
2018-02-27 23:43:35.846 18623 ERROR neutron     self.etcd.start()
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/networking_calico/agent/dhcp_agent.py", line 167, in start
2018-02-27 23:43:35.846 18623 ERROR neutron     super(CalicoEtcdWatcher, self).start()
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/networking_calico/etcdutils.py", line 133, in start
2018-02-27 23:43:35.846 18623 ERROR neutron     cluster_id, last_revision = etcdv3.get_status()
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/networking_calico/etcdv3.py", line 268, in get_status
2018-02-27 23:43:35.846 18623 ERROR neutron     status = client.status()
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/etcd3gw/client.py", line 97, in status
2018-02-27 23:43:35.846 18623 ERROR neutron     json={})
2018-02-27 23:43:35.846 18623 ERROR neutron   File "/usr/lib/python2.7/site-packages/etcd3gw/client.py", line 88, in post
2018-02-27 23:43:35.846 18623 ERROR neutron     raise exceptions.ConnectionFailedError(six.text_type(ex))
2018-02-27 23:43:35.846 18623 ERROR neutron ConnectionFailedError: HTTPConnectionPool(host='192.168.8.2', port=2379): Max retries exceeded with url: /v3alpha/maintenance/status (Caused by NewConnectionError('<requests.packages.urllib3.connection.HTTPConnection object at 0x64d1c90>: Failed to establish a new connection: [Errno 111] ECONNREFUSED',))

2. Rewrite election key so non-masters see progress

The problem here was non-masters watching for events on the election key
and seeing nothing for an interval of twice the expected TTL.  With
etcdv2 the master would refresh by rewriting the key, which meant that
non-masters saw an event.  With etcdv3 the idiom is not to rewrite the
key but instead just refresh the related lease object - but the
semantics of that is (intentionally) that the watchers don't see any
event when the lease is refreshed.  We need the non-master watchers to
see an event, so rewrite the key as well as refreshing the lease.

3. Return None value(s) for corrupt v3 data

We have a test that corrupts the etcd data for a WorkloadEndpoint and
expects networking-calico to fix it again when it resyncs from the
Neutron DB.  This test was failing because the resync was failing to
rewrite the correct data, because of incorrectly expecting that they
would be no existing value at that time.

4. Avoid ERROR logs when election key is deleted

We have a test that deletes the master election key in order to force a
new election.  The code to handle this was working fine but dropping
some ERROR logs that ought to have been avoided.

Change-Id: I9313bf2b3de2d6bb7cf1e60e78adcb161474fb66