Disambiguate DHCP agent's subnet lookup for an endpoint
We've been seeing flakes in the Tempest test test_cross_tenant_traffic,
and this commit is the fix for those flakes.  That test includes
launching instances for different tenants, and it commonly happens that
two of those instances - on different Neutron networks - get assigned
the same IP address.  This is technically outside Calico's scope -
because Calico requires that all tenants share a single address space,
and hence it should not be possible for the same IP address to be
allocated to more than instance, even if they belong to different
tenants - however the fix is quite straightforward, and this test used
to pass reliably before the transition to etcdv3, so I think it's worth
fixing anyway.

In order to provide correct DHCP service, the DHCP agent needs to map
each local endpoint IP correctly to a Neutron subnet and network.  Since
the etcdv3 transition, it does that by comparing each endpoint IP
against the CIDR of each known Neutron subnet.  However this is
ambiguous if there are multiple subnets with overlapping CIDRs.  The
result is that the DHCP agent may try to pass multiple IP/MAC mappings,
for the same IP, to a dnsmasq process; the dnsmasq process then
complains about a 'duplicate' and ignores one of the mappings, which
means that that VM will not get DHCP service when it boots, and won't
learn its IP address.

To fix this, we add an annotation to each WorkloadEndpoint with the
corresponding Neutron network_id.  Then, when finding the right subnet
for an endpoint IP, we also check that subnet's network_id against the
network_id for the endpoint.

Change-Id: Iac19e151177d25d26597610008fe94b23d01483a