Initialize privsep infrastructure for Calico DHCP agent
Also handle removal just now of neutron.common.topics.  That now has to
be neutron_lib.agent.topics.

Also enhance the code in our gate hook that emits Calico diags, so that
it doesn't barf itself if something else goes wrong so as to prevent the
whole test from running.

Also remove test_cross_tenant_traffic from the set of Tempest tests that
we run.  This test has been occasionally flaky for a long time, but now
it seems to have increased in failure rate to about 50%.  It has always
been outside the scope that Calico for OpenStack covers, however, so
seems sensible now to just remove it, given that the increased failure
rate is impacting our ability to get good networking-calico changes
merged.

(Technical explanation: This test creates one VM each for two tenants
and tests connectivity between those, subject to appropriate SG config.
The problem for Calico is that both of those VMs get the same IP address
- which Calico does not support because it requires all endpoints to
share an address space.

Specifically what then goes wrong is that one of the VMs does not get
the SSH key metadata that it should; its console log includes:

  ### /etc/init.d/sshd start
  Top of dropbear init script
  Starting dropbear sshd: remove-dropbear-host-keys already run per instance
  WARN: generating key of type ecdsa failed!
  FAIL

I think this is because correct metadata operation with overlapping IPs
requires use of the Neutron metadata agent, and we don't run that in our
DevStack CI setup; instead we run the Nova metadata agent, which is
simpler but only for a single shared address space.

Lack of that key metadata means that a later SSH into that VM fails.

FWIW, other aspects of Calico surprisingly _do_ work up.  (1) The DHCP
agent runs separate dnsmasq instances per tenant, and it's fine for each
of those to hand out the same IP address.  (2) It's ambiguous which VM
the compute host will route to for <the shared fixed IP>, but in this
particular test and setup that doesn't matter because the test maps a
floating IP to the target VM and uses that for SSH, and Felix also
programs a route for <floating IP> to the relevant TAP interface.

So ultimately it's just the SSH key metadata that causes this test to
fail.)

I've had to include _all_ of the above changes in this same commit,
because they are all needed in order for networking-calico CI to pass
again.

Change-Id: Ida567c33bb868e1b4a658fbc3cfc16f3e3b0610d