Initialize privsep infrastructure for Calico DHCP agent
Also handle removal just now of neutron.common.topics.  That now has to
be neutron_lib.agent.topics.

Also enhance the code in our gate hook that emits Calico diags, so that
it doesn't barf itself if something else goes wrong so as to prevent the
whole test from running.

Change-Id: Ida567c33bb868e1b4a658fbc3cfc16f3e3b0610d