Update requirements handling
Do not use -U for requirements, it will update requirements which can
lead to installation of unwanted versions.

Give test-requirements the usual name so that our jobs handle it
properly.

Add neutron to test-requirements so that Depends-On works. OpenStack CI
will install from git.

Update requirements to current version so that requirements check
passes.

Change-Id: I8ed063d024055077aef75cda2acc13af0d90c2bd