Fix watch loops to handle compaction
The etcdutils watcher loop did not respond properly to etcd compaction.
It had an inner loop and an outer loop, and when compaction occurs we
needed to break to the outer loop - so that we reread the current DB
revision - but in fact we only continued in the inner loop, which meant
that we kept restarting an incorrect watch from a revision that no
longer exists.

With etcd3gw as it stands, we can't distinguish between two watch
timeout cases:

1. watch timeout because of genuine inactivity

2. watch timeout because the revision that we asked to start watching
from no longer exists.

For (1) it's OK to restart the watch from the same revision. For (2) it
is not.  But since we can't distinguish them, we have to assume (2) and
reread the current revision - which means breaking to the outer
loop.  And in fact that means that we don't really need the inner loop
at all any more.

Then, as we're rereading the complete status tree every
WATCH_TIMEOUT_SECS (10s), we have to avoid spamming the Neutron DB with
Felix and port status updates that are duplicates of previous updates.
We already had that protection for port statuses; this commit adds
deduplication for agent status as well.

Empirically, the election watcher loop already handled compaction
correctly, but I've reviewed it to understand and document why that is,
and also reduced the timeout for the unlikely windows where compaction
causes a watch to time out instead of delivering the next election key
event.

Change-Id: I48c863388d1a6f54ebcd587dc2944d1b5f585638