Fix watch loops to handle compaction
The etcdutils watcher loop did not respond properly to etcd compaction.
It had an inner loop and an outer loop, and when compaction occurs we
needed to break to the outer loop - so that we reread the current DB
revision - but in fact we only continued in the inner loop, which meant
that we kept restarting an incorrect watch from a revision that no
longer exists.

etcd3gw as it stands cannot distinguish between two watch timeout cases:

1. watch timeout because of genuine inactivity

2. watch timeout because the revision that we asked to start watching
from no longer exists.

However, we could eliminate (1) as a possibility by regularly writing
ourselves into the tree that is being watched; that acts as an elegant
test of round-trip connectivity to the etcd cluster.

Then, if the watch ever times out, it must be because of compaction (2)
or a genuine connectivity problem.  We should handle that by rereading
the current revision - which means breaking to the outer loop.  And then
it turns out that we don't need the inner loop any more at all.

For the status watching in the Neutron driver, that's what we do.  For
the endpoint and subnet watching in the DHCP agent, we let the watch
timeout and then reread the relevant trees, because that is better than
having every compute host write 2 etcd keys every 3 seconds.

Also in this commit:

- Empirically, the election watcher loop already handled compaction
  correctly, but I've reviewed it to understand and document why that
  is, and reduced the timeout for the unlikely windows where compaction
  causes a watch to time out instead of delivering the next election key
  event.

- When we reread the complete status tree, we could generate Felix agent
  status updates to the Neutron DB that are duplicates of previous
  updates.  We already had code to avoid that for port statuses; this
  commit adds deduplication for agent status as well.

- Another example of unnecessary watch-triggered 'duplication' is the
  DHCP agent restarting Dnsmasq when the set of ports that we want it to
  handle, for each Neutron network, is not actually changing at all.
  This commit enhances the code to avoid that.

Change-Id: I48c863388d1a6f54ebcd587dc2944d1b5f585638