Always send high-priority port statuses to Neutron
High priority port status updates happen because Felix
made an update to the status in etcd and Felix only does
that if it saw the port change state.

Since a port may change state outside of our plugin (for
example due to a VM rebuild), assume that any change Felix
spots needs to be reported to Neutron.

Change-Id: I3fe99cdd21f9aeeeef27b124c950560523fb81d1